Remove dependency from Issues to SdkVersionInfo
The `SdkVersionInfo.underlinesToCamelCase()` method is really an
implementation detail of `SdkVersionInfo` and not something for general
use. While it does vaguely do what is needed (with a little bit of
massaging the input data) it makes little sense to use it when the code
is so simple. Removing the dependency will also make it easier to
extract this package into a separate library.

Bug: 293458631
Test: ./gradlew
Change-Id: I70bf42a9fb769cbe0e9ad56cfd874c75b65461d2