Add test to show broken encoding of string literals
Encoding of string literals in the extracted annotations XML file is
incorrect in two ways:
1. It does not escape the string properly.
2. It gets XML encoded twice.

Given a string literal of `'&\"` the correct behavior according to
IntelliJ should be something equivalent to
`&quot;&#39;&amp;\&quot;&quot;`. It does not have to be identical as
there are many ways of encoding XML and escaping Java strings.

However, when that value is currently encoded it is encoded as
`&quot;&amp;apos;&amp;amp;&amp;quot;&quot;` which is displayed in
IntelliJ as `&apos;&amp;&quot;`.

Bug: 293537075
Test: ./gradlew
Change-Id: I08f82621fdfc2a3771e7691c3128e8bb8ded4fdb