Remove doclava1 API to XML conversion options
AFAICT these are no longer used in Android, were never used in Google3
or AndroidX. It is safe to remove the tests as there are duplicates of
those tests that use the `signature-to-jdiff` subcommand in
`SignatureToJDiffCommandTest`.

Submitting in `aosp/main` first (before `aosp/metalava-main`) to ensure
that all uses in all branches downstream from `aosp/main` have been
removed. If there are problems then this can easily be reverted and it
won't prevent other metalava changes being merged across in the
meantime. Once it has passed all post-submits then this will be merged
into `metalava-main`.

Bug: 290825755
Test: ./gradlew
Change-Id: Iecba034282effec7c02fe8d90571494c35f51532