Allow final modifier to be added to methods of unextendable classes
Previously, adding `final` to a class was treated differently depending
on whether the class was instantiable or not. Unfortunately, the same
was not done for methods. This change treats methods the same way.

A couple of existing tests did not provide explicit public constructors
in the released signature files when testing the effect of adding
`final` and so relied on the previous behavior of not being able to add
`final` to non-instantiable classes. Those tests were updated to add the
necessary constructors.

Bug: 294300564
Test: ./gradlew
Change-Id: I9096e97d459c2758c1c29a9d410a40c971e1abc4