Move DriverTest.getAndroidJar() to ...metalava.testing
This change also removes the unnecessary use of `getAndroidJar()` in
the call to `updateGlobalOptionsForTest(...)` in `PsiTypePrinterTest`.
It is unnecessary because that is being used to set the `--classpath`
command line option which is then processed by
`updateGlobalOptionsForTest(...)` to set `options.classpath` which is
then used as the default for the `classpath` parameter of
`parseSources(...)`. However, `PsiTypePrinterTest` explicitly specifies
the `classpath` parameter to a list that it created itself. Hence,
there is no need to pass the `--classpath` command line option.

Bug: 293429095
Test: ./gradlew
Change-Id: I49e62eb6cc27b66f05f0743fac0a15590f9644ad