Detangle MetalavaCommand from main metalava code
Previously, `MetalavaCommand` hard-coded the behavior when no
subcommand was provided to call into `Driver.kt` functions. It also had
a dependency on `Options`. Both of those tightly coupled the command
with the majority of metalava code.

This change modifies `MetalavaCommand` to take `defaultCommand` that is
run when no subcommand is provided on the command line. That is used to
reduce the dependencies of `MetalavaCommand` and also to replace the
`parseOptionsOnly`.

One slight complexity is that the `defaultCommand` needs to be hidden
from the help but it provides options which must be treated as if they
are part of the main `metalava` command as before. Details are in the
comments. To ensure that worked a previous change added an extra test
for what happens when an invalid option is provided to a subcommand.

Bug: 294815714
Test: ./gradlew
Change-Id: I75d8443abb50ac6273887266066c45c86ecf46dd