Stop PsiTypePrinterTest extending DriverTest
Previously, the `PsiTypePrinterTest` extended `DriverTest` so it could
access the `createProject(...)` method. That is generally useful but if
`PsiTypePrinterTest` extends `DriverTest` it would not be able to be
separated out into its own library. So, this change moves that
functionality into a new `...testing.TemporaryFolderOwner` class and
makes both `DriverTest` and `PsiTypePrinterTest` implement it.

Bug: 293429095
Test: ./gradlew
Change-Id: Ia5534f79048e5742e0d53aef57861e7ef64c01e3