Migrate Reporter/IssueConfiguration options to Clikt
Previously, the properties were processed by `Options` but that meant
they were not usable but sub-commands. This change extracts the
properties into a separate `ReporterOptions` class.

In order for Clikt to process all the issue related options correctly
and in order it needs to consume them all at the same time. It also
needs to update the `IssueConfiguration` immediately as it can affect
the behavior of following options. That necessitated adding the
concept of `sideEffectOption` and `compositeSideEffectOption` on top
of Clikt.

FTR, as it is not obvious from the diff, the `setIssueSeverity` and
`setCategorySeverity` methods are moved to overrides of
`ConfigurableAspect.setAspectSeverityForId()`.

Bug: 295013652
Test: ./gradlew
Change-Id: I3e4dd2c574f00661702e93f104afa398239dc347