Partial signature file contains affected subclass definitions
Partial signature files (signature files created with `show-annotation`
inputs) only contain signatures marked with the input annotation.
This restriction deliberately modifies the represented class hierarchy
to exclude the hidden classes, resulting in class definition that do not
necessarily match that in the source.

This becomes problematic when generating stubs from signature files, as
information about the actual class hierarchy is unknown. Also, such
representation does not align with the principles of the partial
signature files, which is required to include all delta between the API
surfaces.

Thus, include the subclass definition affected by marking the class
hidden to the signature files.

Bug: 295032200
Test: ./gradlew and m update-api
Change-Id: I11ee78ba55fe4be87672533e638cd00d210c5d90