Remove unused single-hyphen options
This pattern of single-hyphen options appears to have been inherited
from Doclava although some new options also appear to have been added
in that pattern (probably due to copy-and-pasting existing code). The
single-hyphen options are not listed in the help and do not appear to
be used anywhere that can be searched.

Searching is not comprehensive because not every use will be indexed
and also even if they are some of the strings are quite generic and so
are used in lots of places that do not involve metalava, e.g. when
using doclava. So, one of the purposes of this change is to detect such
cases.

The `-encoding` option is not removed by this as it is used in some
places but it does nothing and will be deprecated and removed in a
follow up.

The `-source` is also not removed by this as it is used in some places
and actually does something but uses should be replaced with
`--java-source`. Cleaning that up will be done in a follow up.

Neither is the `-stubpackages` option for similar reasons.

Bug: 295136054
Test: ./gradlew
(cherry picked from https://android-review.googlesource.com/q/commit:357855837d7e9aa5009396f55afdfb81c27e8bd5)
Merged-In: Ic08ad3681b2a885e0dd655343b5aa655fdc71386
Change-Id: Ic08ad3681b2a885e0dd655343b5aa655fdc71386