Deprecate global options property
Deprecates the options property and annotates existing usages with
`@Suppress("DEPRECATION")` to prevent build time failures. The scope of
the annotation has been kept reasonably small balancing the cost to
annotate 200+ references and the scope of the suppression of the
deprecation.

This also stops calling `Options.parse()` on the global `options` as it
is no longer needed.

Bug: 286023667
Test: ./gradlew
Change-Id: I11c736af7b5e946299f45c366eb6d70972e91097