Switch InternalDescTest to use a signature file
Previously, this used `testCodebase` from the `psi` test package but
that will not be available when that is moved to a separate project.
The solutions were to either rewrite it to read from source files,
which is quite complex, or switch to signature files which are much
simpler to load. The latter is better as this test is not testing the
`PsiBasedCodebase` it is testing the `MethodItem.internalDesc()` method
and so any equivalent `Codebase` will do.

Bug: 287241408
Test: ./gradlew
Change-Id: Ic4be556ddfa781e5c32751d8e3bbdd3451763a88