Extract CLI specific data from OverloadedMethodOrder
Previously, the `OverloadedMethodOrder` was a mixture of the
`Comparator<MethodItem>` and some CLI data like help text. This change
separates them out so that the `OverloadedMethodOrder` can be moved
into the `model.text` package in a follow up.

This uses a special `OptionOverloadedMethodOrder` enum to encapsulate
the mapping from the command line value to the corresponding
`OverloadedMethodOrder` value. An enum was used as there is good
support for using them in metalava options. It was necessary to add
a new `OptionDelegate.map()` method to map from the internal CLI only
enum to the `OverloadedMethodOrder` instance.

Bug: 295311787
Test: ./gradlew
Change-Id: I1a95bf0759413a5bb6d8036fc2bdb0be4793c3f3