Remove non-signature file formats
Previously, the `FileFormat` contains a mixture of signature API based
formats and other files that metalava might create or consume. However,
the only ones that are really needed are the ones for signature files.

There are three main ways that a `FileFormat` can be "input" to
metalava.
1. Provided on the command line via the `--format` option but that only
   supports signature formats.
2. Determined from a file using `parseHeader` but that is only called
   by `ApiFile` and it will fail if it is not a signature file format.
3. Hard coded in metalava, the only use of that is `BASELINE.header()`
   which returns a fixed string.

Apart from that last use (which has been replaced with a simple
constant) all the other uses of non signature file format were just
to handle a `FileFormat` that had been provided as input and as
explained above they never were provided so it was just dead code.

This change makes the following changes to clean up the resulting
code after removing the uses:
* `FileFormat.parseHeader()` is moved to `ApiFile` and the logic
  around calling it has been improved.
* `FileFormatTest` is moved to `SignatureFileParseHeaderTest`,
  the single test method is split into multiple test methods.

Bug: 295311787
Test: ./gradlew
Change-Id: Ie386425c217e1c423cf2392692fbf5e893583ee2