Add SignatureFileFormat for use by SignatureWriter
Previously, `FileFormat` did some of the work to determine what the
format of the signature file would be but it also had to support non
signature files so it was awkward to add lots of signature specific
information. It is also an enum so has a fixed set of values which
could not encapsulate overrides from the command line. As a result
those overriding options had to be passed alongside the `FileFormat`
leading to quite a verbose and repetitive structure.

This change adds `SignatureFileFormat` which is a signature file
specific data class that encapsulates all the information needed by
`SignatureWriter` that affects the format of the signature file.

It contains the following major changes:
* `OverloadedMethodOrder` was moved inside `SignatureFileFormat`.
* Hides all format related propertes from `SignatureOutputOptions` as
  they are all encapsulated in the new `signatureFileFormat` property.
* The latter change required the removal of some code to fail if
  attempting to check compatibility using a `V1` file against a newer
  `V2+` file. That configuration is no longer possible so it could be
  safely removed.

Bug: 295311787
Test: ./gradlew
Change-Id: I5c4625e3e37216dfc79b1763717d36e9f0a5a2d2