Make ApiFile.parseHeader() take a BufferedReader
Allows the removal of the custom `firstLine` in favor of just using
`BufferedReader.readLine()` and will make it easy for a follow up
change to use that to read the format from a file without reading the
whole file.

Added another couple of tests to handle when the file starts with a
lot of blank lines but is not completely blank and when it is
completely blank.

Bug: 295311787
Test: ./gradlew
Change-Id: If6c3f175a9d9d7f229e5000d434af022399500ae