Add --use-same-format-as option to specify signature format
Previously, the only way to change the format of the generated
signature files was by changing the command line parameters which
requires changes to the build system (or the build configuration for
the module being built). Which would often be expensive and/or tedious.

This change adds the`--use-same-format-as` option which will cause
metalava to use the same format as an existing file. Once that has been
added to the command line then updating a single set (api and removed
api) of files from `2.0` to `3.0` would be as simple as modifying the
signature in the `--use-same-format-as` file and then updating the API
as usual.

Bug: 295577788
Test: ./gradlew
Change-Id: I8b3d74662349b9e07f2d92b7d2441f7405e3f251