Allow merging files with different signature formats
Previously, that was disallowed although it is not clear why as the
change was made as part of some work for Android and Android has been
`2.0` for years now. It may just have been caution.

This change removes that because the formats are often just different
representations of the same model, e.g. 2.0 and 3.0 differ only in how
they represent nullability information. So, having a blanket ban seems
too restrictive. If problems arise we can add a more tailored
compatibility check between formats.

Currently, the output format when merging is determined by the command line and default options, not the source files.

Bug: 295577788
Test: ./gradlew
Change-Id: Ie631cfe979383abcd684115b7782186b095110b0