Add merge-signatures subcommand
Adds a subcommand to merge multiple API signature files together. It is
intended to replace the equivalent functionality added to the main
metalava command once uses of that have been migrated to this.

Copies some existing tests from `ApiFileTest` and tags them as
deprecated so that they can be easily identified and removed once the
functionality has been removed from the main metalava command.

Bug: 295737759
Test: ./gradlew
Change-Id: Icad93a4998e15a1890bc9c5bec514c41fff36975