Move non-overridable FileFormat properties into DefaultsVersion
Previously, the `FileFormat` had a mixture of properties, some of which
were fixed by the version and some of which could be overridden. This
change moves the fixed properties into `DefaultsVersion` and makes them
internal only.

This also adds a `defaults: FileFormat` property to `DefaultsVersion`.
That makes it easy to find the defaults against which a `FileFormat`
can be compared.

Bug: 295577788
Test: ./gradlew
Change-Id: Idf253412a12740101e16655b1d3249f3ed73edaa