Remove --previous-api simplify --migrate-nullness
Previously, the `--previous-api` was an alias for `--migrate-nullness`
and the latter could take an optional value. If the value was not
provided then it had no effect. That seems to be the result of
migrating from a situation where nullness migration required two
options, i.e. `--previous-api <file> --migrate-nullness` to one where
it only requires one, i.e. `--migrate-nullness <file>`. That migration
is complete and there are no uses of either `--previous-api` or
`--migrate-nullness` without a value so this change removes the unused
option and simplifies the other.

Bug: 296115567
Test: ./gradlew
Change-Id: I4e0d11b1ff1398a1e7b982caa1343271b7538a1e