Make FileFormat.parseHeader() more efficient
Previously, the `FileFormat.parseHeader` would read whole lines. While
that is safe when reading an actual signature file or an average text
file, it could lead to reading a lot of data unnecessarily if passed a
binary file by accident.

Now, that `FileFormat.V1` is no longer supported the header prefixes of
all the versions are the same making it easier to read and check the
prefix and then read the version. This change reads a fixed number of
characters to look for the signature prefix and then if that matches
there is a high probability that this is a signature file and so safe
to read. If the first few characters are blank then it is a good sign
that the file is at least a text file so should have reasonable length
lines.

Once the version has been read it is used to perform a map lookup
instead of iterating over the `FileFormat` instances as was done
previously.

Bug: 295577788
Test: ./gradlew
Change-Id: I6dbebf147670267f94ccd4fa60c4cb39628ba60c